feat: filelines

+++ src/basic_benchmark/filelines_baseline.cpp
@@ -23,10 +23,11 @@ void filelines_baseline(char* filepath, uint32_t* total_line_num, uint32_t* line
     }
     int cur_len = 0;
     while (1) {
-        int bytes_read = read(handle, bp, BLOCK_SIZE);
+        // 这个地方非常神奇，要是改用int速度就会变慢，不知道为啥:)
+        ssize_t bytes_read = read(handle, bp, BLOCK_SIZE);
         if (bytes_read <= 0)
             break;
-        for (int i = 0; i < bytes_read; i++) {
+        for (ssize_t i = 0; i < bytes_read; i++) {
             if (bp[i] == '\n') {
                 (*total_line_num)++;
                 if (cur_len >= MAX_LEN)
